+++ src/main/java/co/orange/ddanzi/common/domain/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package co.orange.ddanzi.global.common.domain;
+package co.orange.ddanzi.common.domain;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
