+++ src/main/java/co/orange/ddanzi/controller/ItemController.java
@@ -3,8 +3,8 @@ package co.orange.ddanzi.controller;
 import co.orange.ddanzi.domain.user.User;
 import co.orange.ddanzi.dto.item.ConfirmProductRequestDto;
 import co.orange.ddanzi.dto.item.SaveItemRequestDto;
-import co.orange.ddanzi.global.common.error.Error;
-import co.orange.ddanzi.global.common.response.ApiResponse;
+import co.orange.ddanzi.common.error.Error;
+import co.orange.ddanzi.common.response.ApiResponse;
 import co.orange.ddanzi.repository.AddressRepository;
 import co.orange.ddanzi.repository.UserRepository;
 import co.orange.ddanzi.service.ItemService;
