+++ src/main/java/co/orange/ddanzi/controller/OrderController.java
@@ -1,9 +1,8 @@
 package co.orange.ddanzi.controller;
 
 import co.orange.ddanzi.dto.order.CreateOrderRequestDto;
-import co.orange.ddanzi.global.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.ApiResponse;
 import co.orange.ddanzi.service.OrderService;
-import co.orange.ddanzi.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
