+++ src/main/java/co/orange/ddanzi/controller/SearchController.java
@@ -1,6 +1,6 @@
 package co.orange.ddanzi.controller;
 
-import co.orange.ddanzi.global.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.ApiResponse;
 import co.orange.ddanzi.service.SearchService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
