+++ src/main/java/co/orange/ddanzi/global/jwt/JwtFilter.java
@@ -1,4 +1,4 @@
-package co.orange.ddanzi.global.config.jwt;
+package co.orange.ddanzi.global.jwt;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -8,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
