+++ src/main/java/co/orange/ddanzi/service/HomeService.java
@@ -5,11 +5,11 @@ import co.orange.ddanzi.domain.product.*;
 import co.orange.ddanzi.domain.user.User;
 import co.orange.ddanzi.dto.ProductInfo;
 import co.orange.ddanzi.dto.home.*;
-import co.orange.ddanzi.global.common.error.Error;
-import co.orange.ddanzi.global.common.response.ApiResponse;
-import co.orange.ddanzi.global.common.response.Success;
-import co.orange.ddanzi.global.config.jwt.AuthUtils;
-import co.orange.ddanzi.global.config.redis.RedisRepository;
+import co.orange.ddanzi.common.error.Error;
+import co.orange.ddanzi.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.Success;
+import co.orange.ddanzi.global.jwt.AuthUtils;
+import co.orange.ddanzi.global.redis.RedisRepository;
 import co.orange.ddanzi.repository.*;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
