+++ src/main/java/co/orange/ddanzi/service/InterestService.java
@@ -4,13 +4,12 @@ import co.orange.ddanzi.domain.product.Product;
 import co.orange.ddanzi.domain.user.InterestProduct;
 import co.orange.ddanzi.domain.user.User;
 import co.orange.ddanzi.dto.interest.InterestResponseDto;
-import co.orange.ddanzi.global.common.error.Error;
-import co.orange.ddanzi.global.common.response.ApiResponse;
-import co.orange.ddanzi.global.common.response.Success;
-import co.orange.ddanzi.global.config.jwt.AuthUtils;
+import co.orange.ddanzi.common.error.Error;
+import co.orange.ddanzi.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.Success;
+import co.orange.ddanzi.global.jwt.AuthUtils;
 import co.orange.ddanzi.repository.InterestProductRepository;
 import co.orange.ddanzi.repository.ProductRepository;
-import co.orange.ddanzi.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
