+++ src/main/java/co/orange/ddanzi/service/ItemService.java
@@ -5,9 +5,9 @@ import co.orange.ddanzi.domain.product.Product;
 import co.orange.ddanzi.domain.user.User;
 import co.orange.ddanzi.dto.item.SaveItemRequestDto;
 import co.orange.ddanzi.dto.item.SaveItemResponseDto;
-import co.orange.ddanzi.global.common.error.Error;
-import co.orange.ddanzi.global.common.response.ApiResponse;
-import co.orange.ddanzi.global.common.response.Success;
+import co.orange.ddanzi.common.error.Error;
+import co.orange.ddanzi.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.Success;
 import co.orange.ddanzi.repository.*;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
