+++ src/main/java/co/orange/ddanzi/service/OrderService.java
@@ -8,11 +8,11 @@ import co.orange.ddanzi.domain.user.User;
 import co.orange.ddanzi.dto.AddressInfo;
 import co.orange.ddanzi.dto.order.CheckProductResponseDto;
 import co.orange.ddanzi.dto.order.CreateOrderRequestDto;
-import co.orange.ddanzi.global.common.exception.ItemNotFoundException;
-import co.orange.ddanzi.global.common.exception.ProductNotFoundException;
-import co.orange.ddanzi.global.common.response.ApiResponse;
-import co.orange.ddanzi.global.common.response.Success;
-import co.orange.ddanzi.global.config.jwt.AuthUtils;
+import co.orange.ddanzi.common.exception.ItemNotFoundException;
+import co.orange.ddanzi.common.exception.ProductNotFoundException;
+import co.orange.ddanzi.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.Success;
+import co.orange.ddanzi.global.jwt.AuthUtils;
 import co.orange.ddanzi.repository.AddressRepository;
 import co.orange.ddanzi.repository.DiscountRepository;
 import co.orange.ddanzi.repository.ItemRepository;
