+++ src/main/java/co/orange/ddanzi/service/ProductService.java
@@ -3,16 +3,11 @@ package co.orange.ddanzi.service;
 import co.orange.ddanzi.domain.product.Category;
 import co.orange.ddanzi.domain.product.Discount;
 import co.orange.ddanzi.domain.product.Product;
-import co.orange.ddanzi.domain.user.Address;
-import co.orange.ddanzi.domain.user.User;
-import co.orange.ddanzi.dto.AddressInfo;
 import co.orange.ddanzi.dto.item.ConfirmProductRequestDto;
 import co.orange.ddanzi.dto.item.ConfirmProductResponseDto;
-import co.orange.ddanzi.dto.order.CheckProductResponseDto;
-import co.orange.ddanzi.global.common.error.Error;
-import co.orange.ddanzi.global.common.response.ApiResponse;
-import co.orange.ddanzi.global.common.response.Success;
-import co.orange.ddanzi.global.config.jwt.AuthUtils;
+import co.orange.ddanzi.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.Success;
+import co.orange.ddanzi.global.jwt.AuthUtils;
 import co.orange.ddanzi.repository.*;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
