+++ src/main/java/co/orange/ddanzi/service/auth/AuthService.java
@@ -6,11 +6,11 @@ import co.orange.ddanzi.domain.user.enums.UserStatus;
 import co.orange.ddanzi.dto.auth.AuthResponseDto;
 import co.orange.ddanzi.dto.auth.VerifyRequestDto;
 import co.orange.ddanzi.dto.auth.VerifyResponseDto;
-import co.orange.ddanzi.global.common.error.Error;
-import co.orange.ddanzi.global.common.response.ApiResponse;
-import co.orange.ddanzi.global.common.response.Success;
-import co.orange.ddanzi.global.config.jwt.AuthUtils;
-import co.orange.ddanzi.global.config.jwt.JwtUtils;
+import co.orange.ddanzi.common.error.Error;
+import co.orange.ddanzi.common.response.ApiResponse;
+import co.orange.ddanzi.common.response.Success;
+import co.orange.ddanzi.global.jwt.AuthUtils;
+import co.orange.ddanzi.global.jwt.JwtUtils;
 import co.orange.ddanzi.repository.AuthenticationRepository;
 import co.orange.ddanzi.repository.UserRepository;
 import jakarta.transaction.Transactional;
